containers/ann/requirements.txt
git+git://github.com/safeh2o/swot-python-analysis@packaged
azure-storage-blob
python-dotenv
